requirements.txt
opencv-python~=4.9.0.80
pandas~=2.2.2
streamlit~=1.34.0
torch~=2.2.0
torchinfo~=1.8.0
matplotlib~=3.9.0
numpy~=1.26.4
torchvision~=0.17.0
pillow~=10.2.0
tqdm~=4.66.4
scikit-image~=0.23.2
tensorboard~=2.16.2
facenet-pytorch~=2.6.0
timm~=1.0.3
jupyterlab~=4.2.0